+++ lib/booqable/auth.rb
@@ -58,7 +58,7 @@ module Booqable
         redirect_uri: redirect_uri,
         read_token: read_token,
         write_token: write_token,
-        refresh_token_wrapper: refresh_token_wrapper
+        around_refresh_token: around_refresh_token
       } if oauth_authenticated?
 
       builder.use Booqable::Middleware::Auth::ApiKey, {